9348 close handle

+++ components/tools/OmeroPy/src/omero/gateway/__init__.py
@@ -743,12 +743,15 @@ class BlitzObjectWrapper (object):
 
         # Using omero.cmd.Delete rather than deleteObjects since we need
         # spec/id pairs rather than spec+id_list as arguments
-        handle = self._conn.c.sf.submit(dcs, self._conn.SERVICE_OPTS)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()
+        if len(dcs):
+            doall = omero.cmd.DoAll()
+            doall.requests = dcs
+            handle = self._conn.c.sf.submit(doall, self._conn.SERVICE_OPTS)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()
 
     def removeAnnotations (self, ns):
         """
@@ -764,12 +767,13 @@ class BlitzObjectWrapper (object):
         for al in self._getAnnotationLinks(ns=ns):
             a = al.child
             ids.append(a.id.val)
-        handle = self._conn.deleteObjects('/Annotation', ids)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()        
+        if len(ids):
+            handle = self._conn.deleteObjects('/Annotation', ids)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()        
     
     # findAnnotations(self, ns=[])
     def getAnnotation (self, ns=None):
@@ -6619,7 +6623,12 @@ class _ImageWrapper (BlitzObjectWrapper):
         for chunk in ofw.getFileInChunks():
             outfile.write(chunk)
         outfile.close()
-        self._conn.deleteObjects('/OriginalFile', todel) # No error handling?
+        handle = self._conn.deleteObjects('/OriginalFile', todel)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
+
         return os.path.splitext(f.name.val)[-1], f.mimetype.val
         
     def renderImage (self, z, t, compression=0.9):
@@ -7039,7 +7048,10 @@ class _ImageWrapper (BlitzObjectWrapper):
 
     def _deleteSettings(self):
         handle = self._conn.deleteObjects("/Image/Pixels/RenderingDef", [self.getId()])
-        self._conn._waitOnCmd(handle)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
 
     def _collectRenderOptions (self):
         """
